requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
redis>=5.0.1
celery>=5.3.4
scrapy>=2.11.0
pandas>=2.1.3
numpy>=1.26.2
matplotlib>=3.8.2
plotly>=5.18.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.1
beautifulsoup4>=4.12.2
pytest>=7.4.3
pytest-asyncio>=0.21.1
pydantic>=2.5.1
alembic>=1.12.1
prometheus-client>=0.19.0
grafana-api>=1.0.3
kubernetes>=28.1.0
docker>=6.1.3
httpx>=0.25.1
pyjwt>=2.8.0
bcrypt>=4.0.1
gunicorn>=21.2.0
flower>=2.0.1
pytest-cov>=4.1.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
isort>=5.12.0
pre-commit>=3.5.0
